requirements.txt
anyio==4.9.0
certifi==2025.4.26
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
python-dotenv==1.0.1
python-telegram-bot==20.7
pytz==2025.2
sniffio==1.3.1
typing_extensions==4.13.2
